adding functionality for battles between characters

+++ Task1_RPG_Battles/Interfaces/CharacterAttributes/IPaladinAttributes.cs
@@ -2,7 +2,7 @@ namespace ADS_A1.Interfaces.CharacterAttributes;
 
 public interface IPaladinAttributes : ICharacterAttributes
 {
-    public int Mana { get; }
-    public int MaxMana { get; }
+    public double Mana { get; }
+    public double MaxMana { get; }
     public int HolyPower { get; }
 }
