+++ Task1_RPG_Battles/Objects/Attributes/Builders/CharacterAttributesBuilder.cs
@@ -12,41 +12,41 @@ public class CharacterAttributesBuilder
     // whilst decoupling itself and remaining independant so that it can be improved upon later
     // https://www.reddit.com/r/dotnet/comments/1adwyic/factory_vs_builder/
     
-    private int _health;
-    private int _maxHealth;
-    private int _attack;
-    private int _defense;
-    private int _speed;
+    private double _health;
+    private double _maxHealth;
+    private double _attack;
+    private double _defense;
+    private double _speed;
     private int _level;
-    private int _experience;
-    private int _experienceToNextLevel;
-    private int _gold;
+    private double _experience;
+    private double _experienceToNextLevel;
+    private double _gold;
 
-    public CharacterAttributesBuilder SetHealth(int health)
+    public CharacterAttributesBuilder SetHealth(double health)
     {
         _health = health;
         return this;
     }
 
-    public CharacterAttributesBuilder SetMaxHealth(int maxHealth)
+    public CharacterAttributesBuilder SetMaxHealth(double maxHealth)
     {
         _maxHealth = maxHealth;
         return this;
     }
 
-    public CharacterAttributesBuilder SetAttack(int attack)
+    public CharacterAttributesBuilder SetAttack(double attack)
     {
         _attack = attack;
         return this;
     }
 
-    public CharacterAttributesBuilder SetDefense(int defense)
+    public CharacterAttributesBuilder SetDefense(double defense)
     {
         _defense = defense;
         return this;
     }
 
-    public CharacterAttributesBuilder SetSpeed(int speed)
+    public CharacterAttributesBuilder SetSpeed(double speed)
     {
         _speed = speed;
         return this;
@@ -58,19 +58,19 @@ public class CharacterAttributesBuilder
         return this;
     }
 
-    public CharacterAttributesBuilder SetExperience(int experience)
+    public CharacterAttributesBuilder SetExperience(double experience)
     {
         _experience = experience;
         return this;
     }
 
-    public CharacterAttributesBuilder SetExperienceToNextLevel(int experienceToNextLevel)
+    public CharacterAttributesBuilder SetExperienceToNextLevel(double experienceToNextLevel)
     {
         _experienceToNextLevel = experienceToNextLevel;
         return this;
     }
 
-    public CharacterAttributesBuilder SetGold(int gold)
+    public CharacterAttributesBuilder SetGold(double gold)
     {
         _gold = gold;
         return this;
@@ -97,17 +97,18 @@ public class CharacterAttributesBuilder
         
     }
     
-    // these expression bodies interface the private properties of this object
+    
+    // these expression bodies doubleerface the private properties of this object
     // to provide read only access to these private properties and allowing
     // the object this class is building safe access to these properties
-    // maintaining data integrity of this object
+    // madoubleaining data doubleegrity of this object
     // TODO add a refernce=> further research of builder functions and this implmenetation
-        public int Health => _health;
-        public int Attack => _attack;
-        public int Defense => _defense;
-        public int Speed => _speed;
-        public int Level => _level;
-        public int Experience => _experience;
-        public int ExperienceToNextLevel => _experienceToNextLevel;
-        public int Gold => _gold;
+    public double Health => _health;
+    public double Attack => _attack;
+    public double Defense => _defense;
+    public double Speed => _speed;
+    public int Level => _level;
+    public double Experience => _experience;
+    public double ExperienceToNextLevel => _experienceToNextLevel;
+    public double Gold => _gold;
 }
