+++ Task1_RPG_Battles/Objects/Attributes/PaladinAttributes.cs
@@ -6,8 +6,9 @@ namespace ADS_A1.objects.Attributes;
 
 public class PaladinAttributes : CharacterAttributes, IPaladinAttributes
 {
-    public int Mana { get; }
-    public int MaxMana { get; }
+    public double Mana { get; }
+    public double MaxMana { get; }
+    // future Paladin ability resource:
     public int HolyPower { get; }
 
     public PaladinAttributes(PaladinAttributesBuilder attributes) : base(attributes)
